Rearrange/rewrite stories for storybook 6
