+++ packages/components/src/internal/components/samples/SampleCreationTypeOption.tsx
@@ -1,0 +1,72 @@
+import React, {FC, memo} from "react";
+import classNames from "classnames";
+import {SVGIcon} from "../base/SVGIcon";
+
+export enum CreationType {
+    Derivatives = "Derivatives",
+    PooledSamples = "Pooled Samples",
+    Aliquots = "Aliquots"
+}
+
+export interface CreationTypeModel {
+    type: CreationType,
+    description: string,
+    requiresMultipleParents: boolean,
+    iconSrc?: string,
+    iconUrl?: string
+}
+
+export const CREATION_TYPE_OPTIONS = [
+    {
+        type: CreationType.Derivatives,
+        description: "Create multiple output samples per parent.",
+        requiresMultipleParents: false,
+        iconSrc: 'derivatives'
+    },
+    {
+        type: CreationType.PooledSamples,
+        description: "Put multiple samples into pooled outputs.",
+        requiresMultipleParents: true,
+        iconSrc: "pooled"
+    },
+    {
+        type: CreationType.Aliquots,
+        description: "Create aliquot copies from each parent sample.",
+        requiresMultipleParents: false,
+        iconSrc: "aliquots"
+    }
+]
+
+interface OptionProps {
+    option: CreationTypeModel
+    isSelected: boolean
+    onChoose: (evt) => void
+    showIcon: boolean
+}
+
+// exported only for testing and storybook
+export const SampleCreationTypeOption: FC<OptionProps> = memo(props => {
+    const { option, isSelected, onChoose, showIcon } = props;
+
+    return (
+        <div className={classNames({'creation-type-selected': isSelected})}>
+            {showIcon &&
+            <div className="creation-type-icon">
+                {option.iconUrl && <img src={option.iconUrl} alt={option.type}/>}
+                {option.iconSrc && <SVGIcon iconDir="_images" iconSrc={option.iconSrc}/>}
+            </div>
+            }
+            <label className="creation-type-choice">
+                <input
+                    checked={isSelected}
+                    type="radio"
+                    name="creationType"
+                    value={option.type}
+                    onChange={onChoose}/> {option.type}
+                <div className="creation-type-choice-description">
+                    {option.description}
+                </div>
+            </label>
+        </div>
+    )
+});
